+++ docs/change_log.rst
@@ -56,7 +56,8 @@ The following are only available within Stem's `git repository
   * Support and validation for `ed25519 certificates <api/descriptor/certificate.html>`_ (`spec <https://gitweb.torproject.org/torspec.git/tree/cert-spec.txt>`_, :trac:`21558`)
   * Moved from the deprecated `pycrypto <https://www.dlitz.net/software/pycrypto/>`_ module to `cryptography <https://pypi.python.org/pypi/cryptography>`_ for validating signatures (:trac:`21086`)
   * Sped descriptor reading by ~25% by deferring defaulting when validating
-  * Support for protocol descriptor fields (:spec:`eb4fb3c`)
+  * Added server descriptor's new extra_info_sha256_digest attribute (:spec:`0f03581`)
+  * Added server descriptor's new protocol attribute (:spec:`eb4fb3c`)
   * Shared randomness properties weren't being read in votes (:trac:`21102`)
 
  * **Utilities**